Reset password form in modal and change controllers to device dir
