+++ app/views/devise/passwords/_modal.html.slim
@@ -1,11 +1,9 @@
-.modal.fade id="sessionModal"
+.modal.fade id="passwordModal"
   .modal-dialog
     .modal-content
       .modal-header.bg-light-brown
-        h4.modal-title= 'Iniciar sesión'
+        h4.modal-title= 'Reestablecer contraseña'
         button.btn-close type='button' data-bs-dismiss='modal' aria-label="Close"
       .modal-body
         div.container
           = render "new"
-      .modal-footer
-        = link_to "¿Has olvidado la contraseña?", new_password_path(resource_name), remote: true
