+++ app/views/devise/passwords/_new.html.slim
@@ -1,10 +1,11 @@
-h2 Forgot your password?
 = form_for(resource, as: resource_name, url: password_path(resource_name), html: { method: :post }) do |f|
   = render "devise/shared/error_messages", resource: resource
+  .field
+    = 'Introduce tu dirección de correo electrónico y te enviaremos un enlace para restablecer tu contraseña.'
+    br/
   .field
     = f.label :email
     br/
     = f.email_field :email, autofocus: true, autocomplete: "email"
   .actions
-    = f.submit "Send me reset password instructions"
-= render "devise/shared/links"
+    = f.submit "Reestablecer contraseña"
