+++ app/views/devise/passwords/new.js.erb
@@ -1,0 +1,1 @@
+$('<%= j render "modal", title: "Password" %>').modal('show');
