+++ config/routes.rb
@@ -1,5 +1,10 @@
 Rails.application.routes.draw do
-  devise_for :users, controllers: { sessions: 'my_session' }
+  devise_for :users, {
+    controllers: {
+      sessions: 'devise/my_session',
+      passwords: 'devise/my_passwords'
+    }
+  }
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
   get 'home/index'
   root to: 'home#index'
